Added support for depth-varying Kz
This involved making some single-value inputs into strings which get parsed as lists of numbers.  There's also an interaction between the depth and Kz field to encourage use of matching list lengths.  I fixed some but not all of the cases where a value is hardwired in two different places, asking for trouble.